Password reprompt (#1365)
* Make card number hidden

* Add support for password reprompt

* Rename PasswordPrompt to Reprompt

* Protect autofill

* Use Enums.CipherRepromptType

* Fix iOS not building

* Protect iOS autofill

* Update to match jslib

* Fix failing build

+++ src/App/Pages/Vault/ViewPage.xaml.cs
@@ -128,6 +128,10 @@ namespace Bit.App.Pages
                 }
                 else
                 {
+                    if (!await _vm.PromptPasswordAsync())
+                    {
+                        return;
+                    }
                     await Navigation.PushModalAsync(new NavigationPage(new AddEditPage(_vm.CipherId)));
                 }
             }
@@ -142,6 +146,10 @@ namespace Bit.App.Pages
         {
             if (DoOnce())
             {
+                if (!await _vm.PromptPasswordAsync())
+                {
+                    return;
+                }
                 var page = new AttachmentsPage(_vm.CipherId);
                 await Navigation.PushModalAsync(new NavigationPage(page));
             }
@@ -151,6 +159,10 @@ namespace Bit.App.Pages
         {
             if (DoOnce())
             {
+                if (!await _vm.PromptPasswordAsync())
+                {
+                    return;
+                }
                 var page = new SharePage(_vm.CipherId);
                 await Navigation.PushModalAsync(new NavigationPage(page));
             }
@@ -160,6 +172,10 @@ namespace Bit.App.Pages
         {
             if (DoOnce())
             {
+                if (!await _vm.PromptPasswordAsync())
+                {
+                    return;
+                }
                 if (await _vm.DeleteAsync())
                 {
                     await Navigation.PopModalAsync();
@@ -171,6 +187,10 @@ namespace Bit.App.Pages
         {
             if (DoOnce())
             {
+                if (!await _vm.PromptPasswordAsync())
+                {
+                    return;
+                }
                 var page = new CollectionsPage(_vm.CipherId);
                 await Navigation.PushModalAsync(new NavigationPage(page));
             }
@@ -180,6 +200,10 @@ namespace Bit.App.Pages
         {
             if (DoOnce())
             {
+                if (!await _vm.PromptPasswordAsync())
+                {
+                    return;
+                }
                 var page = new AddEditPage(_vm.CipherId, cloneMode: true, viewPage: this);
                 await Navigation.PushModalAsync(new NavigationPage(page));
             }
